[gap-decorations] Perf tests for GapDecorations
This CL introduces perf tests for the CSSGapDecoration feature that is
currently experimental.

Since the feature includes work on both layout and paint, we added tests
in both of these areas for all the currently supported layout types.

Important to note that the tests are for the most part mirrors of
already existing perf tests, but with the addition of GapDecorations.

Bug: 357648037
Change-Id: If34a9b2b2dfa3748620f9a1cf229b1c9a2d3543b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7185447
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1548820}