[Chrome Next] Remove test that is no longer relevant
This eliminates an ancient test that ensured that the bookmarks bar
was explicitly set to not-visible when it was unparented from the
browser during shutdown.

This test is blocking migration to the new layout to support toolbar-
height side panel.

However, in modern Views, an unparented view is not rendered at all.
Furthermore, its visibility would no longer be set by the layout, as
it would have no effect. And even if it were, whether it got updated
would depend on the layout running, which the test did not ensure.
This is probably why it was already flaky on a number of platforms.

Bug: 40097152
Change-Id: I74778da9c346a0ec3188e0f245f0af646668bfe6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7186239
Reviewed-by: David Pennington <dpenning@chromium.org>
Auto-Submit: Dana Fried <dfried@chromium.org>
Commit-Queue: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1548902}