[iOS]Signin coordinator completion checks the cordinator’s value
One hypothesis for the cause of crbug.com/460078501 is that some some
signin completion cause the wrong signin coordinator to be stopped. This
is usually avoided in coordinator because informing the owner that the
coordinator can be stopped is usually done through the delegate, giving
the coordinator as argument, and the owner can check that it received
the expected coordinator. This CL ensure that signin completion can
provide the same safety.

This CL was almost entirely written by Gemini.

Bug: 460078501
Change-Id: Id524f0a2e53ce9ed783b1ece8dd5d6a25018fc14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7170519
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Auto-Submit: Arthur Milchior <arthurmilchior@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549093}