[ClankNavCap] Rework WindowOpenDisposition into TabLaunchType
The current external navigation logic takes into account the
WindowOpenDisposition defined for a WebContents instance when it is
created.

WOD is used to influence some navigation outcome: when
NEW_FOREGROUND_TAB we are sure the tab was created and presented to the
user.

WOD on W/M/L is used to also detect other scenarios such as Web Popup or
user-modified navigations (e.g. shift-click, middle-click). This is not
the case in Clank where WOD is converted into TabLaunchType (main
example at
org.chromium.chrome.browser.tabmodel.TabModelJniBridge.openNewTab) which
is more widespread in the codebase.

This means that in some cases, the WOD may not be set in Clank because
in the WebContents creation flow this information is not provided.

Rather than changing the code to push WOD into WebContents, it could
make more sense to adopt TabLaunchType in the external navigation logic.
TabLaunchType als captures Android specific interactions that are not
present in WOD.

Bug: 375123758
Change-Id: I5a46a8726881dba5cc5cfe9b114458d5230afe51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7177281
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Fabio Muratori <muratori@google.com>
Cr-Commit-Position: refs/heads/main@{#1549096}