LPP: Return SamlConfirmPasswordScreen results via the WizardController
This allows us to extend this in the future and differentiate between
initial setup and reauthentication. As per go/gb-pin-policy-dd the plan
is to move the SamlConfirmPasswordScreen to after the cryptohome mount
for initial login to avoid asking the user for their password twice
during initial setup.

Bug: b:445665662
Change-Id: I172f191c5825765582065a1dc360a234a0d0d1e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7169067
Reviewed-by: Osama Fathy <osamafathy@google.com>
Reviewed-by: Ivan Šandrk <isandrk@chromium.org>
Commit-Queue: Fahad Mansoor <fahadmansoor@google.com>
Reviewed-by: Leon Masopust <lmasopust@google.com>
Cr-Commit-Position: refs/heads/main@{#1549103}