Use unconnected RuleSets for duplicate contents with cascade layers
We currently deduplicate shared StyleSheetContents during RuleSet-
building in order to avoid "collapsing" distinct anonymous layers
into one. This is done in a somewhat roundabout way with a "fake"
rule mutation event, but is otherwise not an unreasonable way
of handling it.

However, it does not handle the case where the same CSSStyleSheet
instance appears in the list of active stylesheets (which can happen
when the same sheet is listed multiple times in adoptedStyleSheets).
The exact behavior is not handled explicitly by specs, but what we're
doing is not a reasonable default: other things in the RuleSet
(e.g. style rules, named constructs defined by at-rules) do "see"
the different orders when the same sheet is specified multiple times
(see examples in Issue 462744687). Layers are just inconsistent with
this.

To address this, this CL tracks which StyleSheetContents we've already
seen (shared or otherwise), and adds an *unconnected* RuleSet for any
already-seen contents. (Unconnected RuleSets are one-off RuleSets that
ignore the RuleSet cache on StyleSheetContents). In sharing cases
(multiple CSSStyleSheets, same StyleSheetContents) this means we'll
cache less, but also deduplicate less. Under the assumption that
adding the same StyleSheetContents more than once is a rare corner case,
the performance characteristics of this seem unimportant anyway.

This change aligns our behavior with Firefox and Safari.

Fixed: 462744687
Change-Id: I564f7598204227213bce5b9b8eeb62c77acd079b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7185810
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549143}