[WebAuthn] Refactor CredentialSorter for use on Android
`MechanismSorter` contains logic for sorting passkeys and passwords
for display in WebAuthn Immediate mode. The same logic should be
applied to credentials being displayed in Android on the Touch To Fill
bottom sheet.

This change adds a `SortableCredential` interface that contains
a Mechanism on desktop and a Touch To Fill credential on Android. The
logic in `MechanismSorter` has been moved to the new
`CredentialSorter` which sorts the generic type. `MechanismSorter` and
the new `TouchToFillCredentialSorter` are wrappers that do the type
conversion.

Android passkeys are treated as GPM passkeys (Enclave on desktop)
because there is no way to distinguish between them.

A forthcoming CL will pass the credentials to
`TouchToFillCredentialSorter` before they are displayed in the TTF
sheet.

NO_IFTTT=Renaming and moving the enums

Bug: 433543129
Change-Id: I07d6d8690425afbbc32b1f9fe6420602ffef1c8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7169985
Reviewed-by: Adem Derinel <derinel@google.com>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549168}