Add debugging screen for SeedFiles.
Because we're moving the seed and other fields to the Seed Files, the
content will not be available from chrome://local-state for debugging.

This change adds the information stored in the Seed Files in the already
existing chrome://metrics-internals/#variations page.

To handle very large values, if the string is larger than 64 characters,
it will be trimmed and a button for copying the value into the clipboard
will be displayed.

The change is visible in the attachment in the bug
https://g-issues.chromium.org/issues/452071822

Bug: 452071822
Change-Id: I12887e615607323da58d16be4f2486b37e89118b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7100841
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549170}