[SyncTests] Cleanup/modernization in SyncServiceImplHarness
SyncServiceImplHarness has a bunch of helpers that (a) are specific to
Sync-the-feature, (b) sometimes do multiple things, and (c) are
misleadingly named.

E.g. EnableSyncForType(type): If Sync-the-feature isn't enabled yet,
this will turn it on, for only the specific type. If it is already
enabled, then it'll enable this type in addition to any that are
already enabled.

This adds new, less confusing helpers (not used yet, but will be soon)
and also adds some comments and TODOs (to migrate away from the old
confusing helpers).

Bug: 353425612
Change-Id: I6afb1994c62d769f1833af5b99b250396a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7197888
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Mahmoud Rashad <mmrashad@google.com>
Cr-Commit-Position: refs/heads/main@{#1549175}