Implement DevTools remote debugging approval mode
This change introduces a new mode for starting the DevTools remote
debugging server, controlled by user preferences and a policy. The
server can now be started in "approval mode" via the chrome://inspect
page, which is reflected in the `kDevToolsRemoteDebuggingEnabled`
preference. The server will only start if the
`kDevToolsRemoteDebuggingAllowed` policy is not set to false. The
server's lifecycle is managed by observing changes to the
`kDevToolsRemoteDebuggingEnabled` preference. The implementation is made
testable by introducing virtual methods for starting and stopping the
HTTP server and pipe handler.

Bug: 460665929
Change-Id: Ib743167c0f6aacf510e420e5159f14fe49ca7cff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7183381
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549203}