policy: base/hash/sha1 -> crypto/obsolete/sha1
Required for backward compatibility with the client_id derived from
/etc/machine-id, which is used as a filename for storing the DM token.

DM tokens are associated by hashed client IDs, and are stored on-disk.
Changing the hashing algorithm for the client id may break already
existing enrollments or create twin enrollments (that is, enrollments
with different client information, but the same DM token). This could
lead to issues with fetching policies and providing browser status
reports. Because of this, this change migrates to the non-deprecated
API.

This should be a mechanical change, with no expected behavior change.

Bug: 430344933
Change-Id: If847524a2c8bf30f5e251d07b77fd79d429cf561
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7142339
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Commit-Queue: Antonio Alphonse <alphonsea@google.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Reviewed-by: Igor <igorcov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549215}