[iOS][CX] Skip csv export test with credential exchange enabled
The CSV export flow should not be tested from password manager EG test
suite with credential exchange enabled, but rather in the credential
export suite, from where it will be possible to use it.

This CL temporarily skips the test with credential exchange enabled, the
test though should actually be moved. TODO is added for that.

Bug: 463313017
Change-Id: Iccd37dcd6de5b29fd2cf67d99f8592a6038ebc3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7198080
Reviewed-by: Sourav Uttam Sinha <sinhasourav@google.com>
Commit-Queue: Rafał Godlewski <rgod@google.com>
Cr-Commit-Position: refs/heads/main@{#1549219}