Fix LastResortAttemptsUntilStaticMatch
https://chromium-review.googlesource.com/c/chromium/src/+/6917375
introduced some new fallback paths, but didn't update the metric to
account for them. This CL does so, and rebuckets the histogram to have
enough buckets to represent the new values.

For reference, on Android, 0.01% of samples reach the end of this
function without finding a fallback. That value is 0.0005% on Linux and
0.0001% on Windows.

OBSOLETE_HISTOGRAM[LastResortAttemptsUntilStaticMatch]=Replaced by LastResortAttemptsUntilStaticMatch2

Bug: 433293142
Change-Id: Iab23f71666071c77e33cb5e7b777181f2839aba1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7186382
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549228}