[Animated PB] Update progress bar from a single source
How the progress bar gets updated: The progress bar's model is updated
by TopToolbarOverlayMediator (TTOM). The values for the progress bar's
model is calculated by ClipDrawableProgressBar (CDPB), and is returned
in a DrawingInfo object.

Race condition: TTOM listens to onLoadProgressChanged, and reacts by
getting the DrawingInfo from CDPB and updating the model. However, this
DrawingInfo might not be up to date, because CDPB also listens to
onLoadProgressChanged and updates the state for calculating the
DrawingInfo. There isn't any ordering guarantees, so if TTOM reacts to
onLoadProgressChanged first, it will get stale DrawingInfo before CDPB
had a chance to update its state.

Delayed update: LoadProgressMediator (LPM) is the source of truth of
whether or not the progress bar is visible. LPM listens to navigation
and loading events and sets the visibility in CDPB accordingly, but TTOM
doesn't listen to these, and the visibility updates aren't propagated to
TTOM either. If there's a long delay between navigation start and the
next onLoadProgressChanged, the TTOM won't trigger a frame and we won't
see a progress bar.

This CL makes TTOM listen to progress updates coming from CDPB after
CDPB has updated its internal state. This fixes the order and guarantees
that the DrawingInfo will be up to date before TTOM gets it.

Bug: 462424894, 447153080
Change-Id: I335df68dad7881cffac247dfa325b7636a8365bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7178162
Commit-Queue: Peilin Wang <peilinwang@google.com>
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549236}