[a11y] Remove kRowGroup from NOTREACHED list in `GetRoleDescription`.
This CL is a partial revert of crrev.com/c/7124361 which marked
`kRowGroup` and a two other roles as NOTREACHED on Android.
The change resulted in crashes for 'kRowGroup` and hence is reverted
for this role onle.

Further investigation will be done to add a test case. The role is now
omitted in all test cases on Android expectations.

AX-Relnotes: n/a
Bug: 460866659, 447360631
Change-Id: Ic32a58a8c4410b2c1d005e265717ea934a06add7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7198359
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Auto-Submit: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549261}