[AIM] Adjust coloring of the request type button.
Most pieces of the button had coloring changes, as well as the dashed
outline style was removed. This more closely matches the mocks.

Some care must be taken to work around the button's normal
setCompoundDrawableTintList or setting a drawableTint in the layout
file, as we need divergent tints on the various drawables, instead of a
single shared one for both.

Bug: 461544172
Change-Id: I9c303d595eef0be5bb9e21b810014652845c0133
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7164018
Commit-Queue: Sky Malice <skym@chromium.org>
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1549295}