[BNPL][Clank][Part4] Handle user clicking BNPL chip before AE completes
This CL handles the user flow for when the BNPL chip is clicked before
the transaction amount has been extracted.

When a user clicks the chip, Java calls the native side to display a
progress screen, as amount extraction is not yet complete. Once
amount extraction finishes and the amount is not supported by any
issuer, the native side calls back to Java with the BnplIssuerContext
to display the BNPL issuer selection screen with all issuers grayed
Out.

In this CL, when all issuers are grayed out, if the user hits the Back
button and goes to the home page, the BNPL chip is grayed out as well,
with the message: “Unavailable for this purchase.”

Video:
http://screencast/cast/NjEyMTc3MDE4Nzc1MTQyNHxkYjkyMzUzNi1mNg

Design: go/autofill-clank-bnpl

Bug: b:438784412, 430575808
Change-Id: I93cef8c4180de51cac96fa17205bda58d8e666ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7158394
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Commit-Queue: Viplav Kadam <viplavkadam@google.com>
Cr-Commit-Position: refs/heads/main@{#1549360}