[blink] Unblock touch moves before dispatching to renderer main
Currently, any touch moves sent to the renderer process as blocking are
only be unblocked on the renderer main thread. But, if the renderer main
thread has determined that touch moves for a given touch sequence can be
unblocked, the compositor thread should dispatch them as non-blocking to
prevent jank due to delayed handling on the main thread.

Bug: 420477114
Change-Id: Icff5357edf60642e09f7b2ff69df01fc5bfc02d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7163795
Commit-Queue: Anand Ravi <anandrv@google.com>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549379}