[Vertical Tabs] Alert indicator button
In order to share the alert indicator button between horizontal and
vertical tabs, create a common interface ParentTab. Due to inheritance
issues it's hard to know that it is also a views::View, so the DCHECK
that the parent tab is the parent in the view hierarchy is removed, and
the color provider will just be taken from the alert indicator button
because it should be the same one as its parent's.

https://screencast.googleplex.com/cast/NjU0MTEzNzU4OTMwNTM0NHwzMzNlOTQ5Ni1mMg

Fixed: 457525173
Change-Id: I2ef413e4f61d00bff2c4400a6e262e89b0c70125
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7182020
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Charles Meng <charlesmeng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549443}