Restore federated session keys on use
If a site tries to register a federated session, but the provider's
session has not yet been restored, attempt to restore the key. If
restoration fails, we would have deleted the session on the first
refresh anyway, so we delete it. If restoration succeeds, we can proceed
with federated registration.

Fixed: 462003255
Change-Id: Ia67ff23c6b51a9760d6fcc7d62b6a7f66a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7182022
Reviewed-by: thefrog <thefrog@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549469}