[contextual-tasks][lens] Upload viewport and context data in StartQueryFlow
This CL updates `LensQueryFlowRouter::StartQueryFlow` to upload the
viewport screenshot and tab context data when the Lens overlay is
invoked with Contextual Tasks enabled. It also now notifies the handle
that the session has started.

Not included in this change is sending the interaction requests and
ending the session before opening the contextual tasks panel.

Bug: 462488630
Change-Id: Ie4aec0067b206866a50a6ff5c7ee6e787b448a51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7201188
Reviewed-by: Duncan Mercer <mercerd@google.com>
Commit-Queue: Juan Mojica <juanmojica@google.com>
Cr-Commit-Position: refs/heads/main@{#1549487}