Move BindOkContinuation() out of media::TypedStatus
This method is only used with HlsDemuxerStatus. It has been a couple of
years since https://crrev.com/1224740 added BindOkContinuation() and no
other code has started using it. So assume it will not be widely used
and move it into the one file that uses it.

Change-Id: Ia728c76c07b6c911d3c8f31b4610c97c0b9bfeb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7201426
Reviewed-by: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549512}