Delete unused AssistiveWindowType::kEmojiSuggestion [10/N]
This value is never applied to a window. So all the logic that
depend on this value are dead code.

For the corresponding Mojo value, it is only ever used in an API
from browser process to IME service shared library [1]. So as long
as the Chromium side never sends this value, the corresponding
Mojo value is also dead code.

[1] https://source.chromium.org/chromium/chromium/src/+/main:chromeos/ash/services/ime/public/mojom/input_method.mojom;l=719;drc=58e43bf54a9e8b399bad1c5c684862a0131dc071

Bug: 444093780
Change-Id: Icc176503d572089cef0b52cb2fd1d2dd2c44b5ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7156110
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Bao-Duy Tran <tranbaoduy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549519}