[OpenScreen] [VideoStack] Update test to properly record laptop screen
The Magewell capture cards are stubborn and won't properly use metrics
sent via ffmpeg. I'm filing a ticket with Magewell, but until then the
plan is to gather the metrics directly from the capture card to ensure
the ffmpeg settings are correct to capture the screen.

You can see the failures in the CAS outputs--those will be addressed in
a follow-up, and hopefully the builder will actually go green.

Bug: 455618553
Change-Id: I927f687fbb8174845931a8d06e7fdda55daa8c5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7186754
Reviewed-by: Zijie He <zijiehe@google.com>
Commit-Queue: Bailey Myers-Morgan <bmyersmorgan@google.com>
Cr-Commit-Position: refs/heads/main@{#1549522}