spanification: Clean up unneeded parameters in
url

This was the result of asking gemini to remove all of the redundant
variables and update call sites. I have also gone ahead and updated all
the for loops that the linter complained were not modern c++ (to range
for loops).

The aim of this patch should be no behaviour changes, however well I've
manually reviewed it, since this patch was generated by AI please review
carefully.

These checks have been in the code for 2 stable releases, and M143 is
about to go to Stable, which means this has been fully fatal in
Canary/Dev. All of these should be completely redundant.

This patch was split with `git cl split --target-range 8 32
--max-depth=2 --no-reviewers -d ~/description.txt`

This is part of auto spanification see internal doc about it:
https://docs.google.com/document/u/1/d/1poEhNARfFas2osqPoXGzVZtBgnOQGQtQNUFTA-MlADs/edit?tab=t.0

This CL was uploaded by an experimental version of git cl split
(https://crbug.com/389069356).

Bug: 431824301
Change-Id: I9514089d1c3a4e5a65e698aeb9024b2f2516952a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7201206
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Commit-Queue: Hayato Ito <hayato@chromium.org>
Reviewed-by: Hayato Ito <hayato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549530}